requirements.txt
amqp==5.0.6
aniso8601==8.1.0
attrs==20.3.0
bcrypt==3.2.0
bidict==0.21.2
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
dnspython==1.16.0
docker==4.4.1
dockerfile-parse==1.1.0
eventlet==0.30.0
Flask==1.1.2
flask-restx==0.2.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.2
prompt-toolkit==3.0.18
proxmoxer==1.1.1
pycparser==2.20
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
